Number of attempts added to Login screen.
If the user leaves the username or password field empty, a message appears onscreen to inform them of this.

The credentials are checked to see if they are valid (match the sample username and password).

If  the user credentials are invalid the number of attempts are reduced. Once the attempts reach 0, the login button is disabled.

If the credentials are valid then the user can enter the app and view the Overview screen.