Sign Out function and option moved to overflow menu
Sign out function(signOutUser()) applied within the onOptionsItemSelected(@NonNull MenuItem item) method to sign out the user using Firebase Authentication.
After signing out the app returns to the login screen. This code is in OverviewActivity.java.

Sign out option is added to the overflow menu instead of the top menu as discussed in a team meeting. This change is in overview_menu.xml.